requirements.txt
ccdexplorer-fundamentals
python-dotenv
grpcio
aiohttp
asyncio
grpcio-tools
betterproto==2.0.0b5
chardet
pytest
python-dotenv